+++ src/service/testPage/problemUpload/index.tsx
@@ -1,0 +1,15 @@
+import { printLoadingText } from './printLoadingText';
+import { createShowSolutionsButton } from './createShowSolutionsButton';
+import { uploadCurrentSolution } from './uploadCurrentSolution';
+import { printRequestOfRefresh } from './printRequestOfRefresh';
+import { parsingDomNodeToUpload } from './parsingDomNodeToUpload';
+import { printIsUploadSuccess } from './printIsUploadSuccess';
+
+export {
+  printLoadingText,
+  createShowSolutionsButton,
+  uploadCurrentSolution,
+  printRequestOfRefresh,
+  parsingDomNodeToUpload,
+  printIsUploadSuccess,
+};
